requirements.txt
Flask>=3.0.0
NetworkX>=3.0
pyvis>=0.3.2
blockfrost-python>=0.6.0
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

